Reorganized tests according to functionalities
